refactor: move NetworkRouter into its own file
